project structure optimized
