+++ models/contacts/filePath.js
@@ -1,0 +1,5 @@
+const path = require("path");
+
+const filePath = path.join(__dirname, "../../db/contacts.json");
+
+module.exports = filePath;
