+++ models/contacts/getContactsList.js
@@ -1,0 +1,10 @@
+const fs = require("fs/promises");
+const filePath = require("./filePath");
+
+const getContactsList = async () => {
+	const data = await fs.readFile(filePath);
+	const contacts = JSON.parse(data);
+	return contacts;
+};
+
+module.exports = getContactsList;
