+++ models/contacts/index.js
@@ -1,0 +1,13 @@
+const getContactsList = require("./getContactsList");
+const getContactById = require("./getContactById");
+const addContact = require("./addContact");
+const updateContact = require("./updateContact");
+const removeContact = require("./removeContact");
+
+module.exports = {
+	getContactsList,
+	getContactById,
+	addContact,
+	updateContact,
+	removeContact,
+};
