+++ routes/api/contacts.js
@@ -8,12 +8,12 @@ const {
 	addContact,
 	updateContact,
 	removeContact,
-} = require("../../models/contacts");
+} = require("../../models/contacts/index");
 
 const router = express.Router();
 
 const schema = Joi.object({
-	name: Joi.string().alphanum().min(2).max(30).required(),
+	name: Joi.string().min(2).max(30).required(),
 	email: Joi.string().email().required(),
 	phone: Joi.string()
 		.pattern(
